% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llcont.R
\name{llcont}
\alias{llcont}
\title{Individual Log-Likelihoods}
\usage{
llcont(x, ...)
}
\arguments{
\item{x}{a model object}

\item{\dots}{arguments passed to specific methods}
}
\value{
An object of class \code{numeric} containing individuals' contributions to the log-likelihood.  The sum of these contributions equals the model log-likelihood.
}
\description{
Obtain log-likelihood values associated with individual observations, evaluated at the ML estimates.
}
\details{
This is a S3 generic function.
Currently, the method is defined for \code{lm}, \code{glm}, \code{glm.nb},
\code{clm}, \code{hurdle}, \code{zeroinfl}, \code{mlogit}, \code{nls},
\code{polr}, \code{rlm}, \code{lavaan} objects.
}
\examples{
## Fit gamma glm, check that sum of llcont() equals
## the model loglikelihood:
clotting <- data.frame(u = c(5,10,15,20,30,40,60,80,100),
                       lot1 = c(118,58,42,35,27,25,21,19,18),
                       lot2 = c(69,35,26,21,18,16,13,12,12))
gam1 <- glm(lot1 ~ log(u), data = clotting, family = Gamma)
sum(llcont(gam1))
logLik(gam1)

}

